Move dali.device.Device class to dali.bus, device module conflicted with device package.
