+++ README.rst
@@ -35,38 +35,53 @@ Library structure
 - ``dali``
 
   - ``address`` - Device addressing
-  - ``bus`` - DALI bus object
+
+  - ``bus`` - DALI bus and attached devices
+
   - ``command`` - Command registry, interface to command decoding
+
   - ``compat`` - Compatibility code for Python 2 and 3
-  - ``device`` - Control devices and events from them
+
+  - ``device`` - DALI control devices as defined in IEC 62386
 
     - ``general`` - Commands and events from part 103
 
   - ``driver`` - Objects to communicate with physical DALI gateways or services
 
     - ``base`` - General driver contracts
+
     - ``hasseb`` - Driver for Hasseb DALI Master (needs to be adopted to dali.driver.base API)
+
     - ``tridonic`` - Driver for Tridonic DALI USB
+
     - ``daliserver`` - Driver for https://github.com/onitake/daliserver (needs to be adopted to dali.driver.base API)
 
   - ``exceptions`` - DALI related exceptions
+
   - ``frame`` - Forward and backward frames
-  - ``gear`` - Control gear
+
+  - ``gear`` - DALI control gear as defined in IEC 62386
 
     - ``emergency`` - Commands from part 202
+
     - ``general`` - Commands from part 102
+
     - ``incandescent`` - Commands from part 205
-    - ``fluorescent`` - Commands from part 201 (not yet implemented)
+
     - ``led`` - Commands from part 207
 
 
 Contributors
 ------------
 
-- Stephen Early (Autor)
+- Stephen Early (Author)
+
 - Robert Niederreiter
+
 - Diogo Gomes
+
 - Caiwan
+
 - Boldie
